make valueFactor parameter a runtime argument

+++ src/main/java/minoaner/evaluation/EvaluateFullMatchingWithLabelHeuristicARCS.java
@@ -153,7 +153,7 @@ public class EvaluateFullMatchingWithLabelHeuristicARCS extends BlockingEvaluati
         final float MIN_SUPPORT_THRESHOLD = 0.01f;
         final int N = (args.length >= 8) ? Integer.parseInt(args[7]) : 3; //top-N relations
         System.out.println("N = "+N);
-        final float valueFactor = (args.length >= 9) ? Integer.parseInt(args[8]) : 0.6f; //the weight of values vs neighbors for the rank aggregation (linear combination)
+        final float valueFactor = (args.length >= 9) ? Float.parseFloat(args[8]) : 0.6f; //the weight of values vs neighbors for the rank aggregation (linear combination)
         
         System.out.println("Getting the top K value candidates...");
         CNPARCS cnp = new CNPARCS();
